eventsApp/notification-service/pyproject.toml
[project]
name = "notification-service"
version = "0.1.0"
description = "notification service for eventsApp"
authors = [
    {name = "Jojees",email = "joji@jojees.net"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pika (>=1.3.2,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"